비밀번호 재설정 API 구현

+++ src/main/java/cloneproject/Instagram/service/EmailCodeService.java
@@ -6,8 +6,15 @@ import java.util.Random;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.stereotype.Service;
 
+import cloneproject.Instagram.entity.member.Member;
 import cloneproject.Instagram.entity.redis.EmailCode;
+import cloneproject.Instagram.entity.redis.ResetPasswordCode;
+import cloneproject.Instagram.exception.CantResetPasswordException;
+import cloneproject.Instagram.exception.MemberDoesNotExistException;
+import cloneproject.Instagram.exception.NoConfirmEmailException;
 import cloneproject.Instagram.repository.EmailCodeRedisRepository;
+import cloneproject.Instagram.repository.MemberRepository;
+import cloneproject.Instagram.repository.ResetPasswordCodeRedisRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -15,14 +22,15 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor 
 @Service
 public class EmailCodeService {
-
     
+    private final MemberRepository memberRepository;
     private final EmailCodeRedisRepository emailCodeRedisRepository;
+    private final ResetPasswordCodeRedisRepository resetPasswordCodeRedisRepository;
     private final EmailService emailService;
 
     public boolean sendEmailConfirmationCode(String username, String email){
 
-        String code = createConfirmationCode();
+        String code = createConfirmationCode(6);
 
         EmailCode emailCode = EmailCode.builder()
                                         .username(username)
@@ -40,19 +48,71 @@ public class EmailCodeService {
         return true;
     }
 
-    public Optional<EmailCode> findByUsername(String username){
-        Optional<EmailCode> emailCode = emailCodeRedisRepository.findByUsername(username);
-        return emailCode;
+    public boolean checkEmailCode(String username, String email, String code){
+        Optional<EmailCode> optionalEmailCode = emailCodeRedisRepository.findByUsername(username);
+
+        if(optionalEmailCode.isEmpty()){
+            throw new NoConfirmEmailException();
+        }
+
+        EmailCode emailCode = optionalEmailCode.get();
+
+        if(!emailCode.getCode().equals(code) || !emailCode.getEmail().equals(email)){
+            return false;
+        }
+
+        emailCodeRedisRepository.delete(emailCode);
+        return true;
     }
 
-    public void deleteEmailCode(EmailCode emailCode){
-        emailCodeRedisRepository.delete(emailCode);
+    public boolean sendResetPasswordCode(String username){
+        Member member = memberRepository.findByUsername(username).orElseThrow(MemberDoesNotExistException::new);
+        String code = createConfirmationCode(30);
+
+        ResetPasswordCode resetPasswordCode = ResetPasswordCode.builder()
+            .username(username)
+            .code(code)
+            .build();
+        
+        resetPasswordCodeRedisRepository.save(resetPasswordCode);
+
+        SimpleMailMessage mailMessage = new SimpleMailMessage();
+
+        String emailText = username + "의 인증코드 " + code;
+
+        mailMessage.setTo(member.getEmail());
+        mailMessage.setSubject("비밀번호 재설정 메일입니다");
+        mailMessage.setText(emailText);
+        emailService.sendEmail(mailMessage);
+        
+        return true;
     }
     
-    public String createConfirmationCode(){
+    public boolean checkResetPasswordCode(String username, String code){
+        Optional<ResetPasswordCode> optionalResetPasswordCode = resetPasswordCodeRedisRepository.findByUsername(username);
+
+        if(optionalResetPasswordCode.isEmpty()){
+            throw new CantResetPasswordException();
+        }
+
+        ResetPasswordCode resetPasswordCode = optionalResetPasswordCode.get();
+
+        if(!resetPasswordCode.getCode().equals(code)){
+            return false;
+        }
+        return true;
+    }
+
+    public void deleteResetPasswordCode(String username){
+        ResetPasswordCode resetPasswordCode = resetPasswordCodeRedisRepository.findByUsername(username)
+            .orElseThrow(CantResetPasswordException::new);
+        resetPasswordCodeRedisRepository.delete(resetPasswordCode);
+    }
+
+    public String createConfirmationCode(int length){
         StringBuffer key = new StringBuffer();
         Random random = new Random();
-        for (int i = 0; i < 6; i++) {
+        for (int i = 0; i < length; i++) {
             boolean isAlphabet = random.nextBoolean();
  
             if(isAlphabet){
